+++ tests/integration/targets/ssh_config/tasks/main.yml
@@ -6,7 +6,6 @@
 - name: Install required libs
   pip:
     name:
-      - stormssh
       - 'paramiko<3.0.0'
     state: present
     extra_args: "-c {{ remote_constraints }}"
